requirements.txt
amqp==5.1.1
ansible==6.1.0
ansible-core==2.13.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
authlib>=1.2.0
autobahn==22.12.1
Automat==22.10.0
bcrypt==4.0.1
billiard==3.6.4.0
cachetools==5.2.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
daphne==4.0.0
Deprecated==1.2.13
Django==3.2.12
django-cors-headers==3.13.0
django-health-check==3.17.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
docutils==0.19
drf-spectacular[sidecar]>=0.25.1
flower==1.2.0
google-auth==2.16.0
gunicorn==20.1.0
humanize==4.4.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
IPy==1.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.3.1
kombu==5.2.4
kubernetes==25.3.0
lockfile==0.12.2
MarkupSafe==2.1.1
msgpack==1.0.4
mysql-connector-python==8.0.32
oauthlib==3.2.2
packaging==23.0
paramiko==2.11.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycrypto==2.6.1
pydantic==1.10.5
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==23.0.0
python-daemon==2.3.2
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
redis==4.4.1
requests==2.28.1
requests-oauthlib==1.3.1
resolvelib==0.8.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tornado==6.2
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.4.1
wrapt==1.14.1
zope.interface==5.5.2
